pyproject.toml
[project]
name = "twilio-energy-monitor"
version = "0.1.0"
description = "Backend system that receives messages from Twilio chat, processes them using Gemini AI, and registers energy counter readings to BigQuery"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.121.2",
    "google-cloud-bigquery>=3.38.0",
    "kaleido>=0.2.1",
    "langchain-google-vertexai>=2.0.11",
    "langgraph>=1.0.3",
    "pillow>=11.1.0",
    "plotly>=5.24.0",
    "pydantic-settings>=2.12.0",
    "pytest>=8.3.4",
    "requests>=2.32.3",
    "twilio>=9.8.6",
    "uvicorn>=0.38.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/twilio_energy_monitor"]
